+++ src/main/java/com/chaosbuffalo/mkultra/core/abilities/KanyeCutter.java
@@ -43,7 +43,7 @@ public class KanyeCutter extends PlayerAbility {
     }
 
     @Override
-    public int getManaCost(int currentRank) {
+    public float getManaCost(int currentRank) {
         return 3 + currentRank * 2;
     }
